tweak diff mode
