+++ lib/ace/mode/diff.js
@@ -41,10 +41,11 @@ var oop = require("../lib/oop");
 var TextMode = require("./text").Mode;
 var Tokenizer = require("../tokenizer").Tokenizer;
 var HighlightRules = require("./diff_highlight_rules").DiffHighlightRules;
+var FoldMode = require("./folding/diff").FoldMode;
 
 var Mode = function() {
     this.$tokenizer = new Tokenizer(new HighlightRules().getRules(), "i");
-    this.foldingRules = new FoldMode();
+    this.foldingRules = new FoldMode(["diff", "index", "\\+{3}", "@@|\\*{5}"], "i");
 };
 oop.inherits(Mode, TextMode);
 
@@ -53,39 +54,5 @@ oop.inherits(Mode, TextMode);
 }).call(Mode.prototype);
 
 exports.Mode = Mode;
-
-
-// folding
-var BaseFoldMode = require("./folding/fold_mode").FoldMode;
-var Range = require("../range").Range;
-
-var FoldMode = exports.FoldMode = function() {};
-oop.inherits(FoldMode, BaseFoldMode);
-
-(function() {
-
-    this.foldingStartMarker = /^(diff|index|@@|\+{3}|\*{5})/;
-
-    this.getFoldWidgetRange = function(session, foldStyle, row) {
-        var line = session.getLine(row);
-        var start = {row: row, column: line.length - 2};
-
-        var regList = ["diff", "index", "\\+{3}", "@@|\\*{5}"];
-        for (var i = 1; i <= regList.length; i++) {
-            var re = RegExp("^(" + regList.slice(0, i).join("|") + ")");
-            if (re.test(line))
-                break;
-        }
-
-        for (var l = session.getLength(); ++row < l; ) {
-            line = session.getLine(row);
-            if (re.test(line))
-                break
-        }
-
-        return  Range.fromPoints(start, {row: row - 1, column: line.length});
-    };
-
-}).call(FoldMode.prototype);
 
 });
