+++ lib/ace/mode/diff_highlight_rules.js
@@ -89,10 +89,10 @@ var DiffHighlightRules = function() {
                     "invalid"
                 ],
             }, {
-                "regex": "^(diff)( .+)$",
-                "token": ["keyword", "string.regex"]
+                "regex": "^(diff)(\\s+--\\w+)?(.+?)( .+)?$",
+                "token": ["variable", "variable", "keyword", "variable"]
             }, {
-                "regex": "^Index .+$",
+                "regex": "^Index.+$",
                 "token": "variable"
             }, {
                 "regex": "^(.*?)(\\s*)$",
